chore: renamed db_util.rs to config.rs
