+++ src/handlers/mod.rs
@@ -1,4 +1,4 @@
-mod db_util;
+mod config;
 pub mod health_check;
 pub mod store_subscribed;
 pub mod subscriptions;
